+++ azure/service.ts
@@ -109,7 +109,7 @@ interface AzureContainer {
     image: pulumi.Input<string>;
     memory: pulumi.Input<number>;
     name: pulumi.Input<string>;
-    port?: pulumi.Input<number>;
+    ports?: azure.types.input.containerservice.GroupContainer["ports"];
     protocol?: pulumi.Input<string>;
     volumes?: pulumi.Input<pulumi.Input<{
         mountPath: pulumi.Input<string>;
@@ -213,7 +213,9 @@ function createGroup(
             name: qualifiedName,
             cpu: pulumi.output(container.cpu).apply(c => c === undefined ? 1 : c),
             memory: memoryInGB,
-            port: targetPortNumber,
+            ports: [{
+                port: targetPortNumber,
+            }],
             protocol: protocol,
             image: imageOptions.image,
             environmentVariables: imageOptions.environment,